warp_kinematics: pass explicit dtypes to from_numpy (parallel IK is dead on warp 1.15)
Upstream pinned warp-lang<1.15.0 with the reason "ik tests fail", then fixed
the incompatibility. This fork has neither the pin nor the fix, and warp 1.15.0
is what a fresh install resolves -- warp-lang arrives transitively via
mujoco-warp (>=1.11.0) and there is no lockfile, so this is almost certainly a
migration artefact rather than a long-standing defect. Worth confirming what
the previous machine had installed.

warp 1.15.0 added an element-size equality check to wp.copy (context.py:12661);
1.14.0 has no such check and the identical call succeeds with correct values.
from_numpy's inference is unchanged across 1.14/1.15/1.17: a (B, N) array
infers vector(length=N), so a 1-D array of B elements of 4N bytes meets a
destination of 4. Still present in 1.17.0.dev.

mlspaces_tests/component_tests/test_kinematics.py: 52 failed, 20 passed before;
72 passed after. Every SimpleWarpKinematics.fk() and .ik() raised on its first
wp.copy, on both cpu and cuda:0.

Neither pixels nor judge_success can be perturbed by this. The RuntimeError is
raised inside wp.copy before any kernel writes, so data.qpos retains its prior
contents and there is no path where IK returns a plausible-but-wrong result.
The symptom is 100% episode loss with tracebacks on any warp-IK path
(grasp_sample.py:133, base_object_manipulation_planner_policy.py:427,:551 --
none catch it; the nearest handler drops the episode), not a shifted camera.

The float(0.0) -> wp.float32(0.0) half of upstream's fix has *zero* numerical
effect -- bitwise identical on warp 1.14.0/1.15.0/1.17.0.dev, cpu and cuda,
across cholesky_solve6, lm_step and mask_jacobian. warp's generated header
defines float(x) as cast_float(x), a C++ 32-bit float; it never meant float64.
It is applied anyway because ruff UP018 flags float(0.0), UP is in this repo's
lint selection, and pre-commit runs ruff-check --fix -- whose autofix rewrites
it to 0.0, which warp treats as a compile-time constant and rejects with
"Error mutating a constant s inside a dynamic loop". One pre-commit run would
otherwise break the file.

One deviation from upstream's enumerated hunks, flagged for review: line 155's
float(mask[i,k]) is the same idiom and is converted too (verified bit-exact).

No pin added. With the fix the code works on 1.14.0, 1.15.0 and 1.17.0.dev;
upstream's warp-lang<1.15.0 was the stop-gap its own later commit superseded,
and a pin that must be bumped by hand is strictly worse than a 12-second test
that fails loudly.

17 tests, ~12 s, no assets or GPU required, parametrised over cpu and cuda:0.
They import the repo's own kernels rather than copies, so a warp bump that
changes them is caught. Includes a bit-exact analytic pin (power-of-two
diagonal, exactly representable and device-independent), a float64 comparison
with ~25x margin (tolerant of FMA contraction, not of a real regression), and
two source guards. Both guards were verified to fire by reintroducing the bug.

+++ molmo_spaces/kinematics/parallel/warp_kinematics.py
@@ -152,7 +152,7 @@ def mask_jacobian(
     i = wp.tid()
     for j in range(3):
         for k in range(nv):
-            J[i, j, k] = J[i, j, k] * float(mask[i, k])
+            J[i, j, k] = J[i, j, k] * wp.float32(mask[i, k])
 
 
 mat66f = wp.types.matrix(shape=(6, 6), dtype=wp.float32)
@@ -166,7 +166,7 @@ def cholesky_solve6(H: mat66f, b: vec6f) -> vec6f:
     L = mat66f()
     for i in range(6):
         for j in range(i + 1):
-            s = float(0.0)
+            s = wp.float32(0.0)
             for k in range(j):
                 s += L[i, k] * L[j, k]
             if i == j:
@@ -177,7 +177,7 @@ def cholesky_solve6(H: mat66f, b: vec6f) -> vec6f:
     # Forward substitution: L @ y = b
     y = vec6f()
     for i in range(6):
-        s = float(0.0)
+        s = wp.float32(0.0)
         for k in range(i):
             s += L[i, k] * y[k]
         y[i] = (b[i] - s) / L[i, i]
@@ -185,7 +185,7 @@ def cholesky_solve6(H: mat66f, b: vec6f) -> vec6f:
     # Backward substitution: L^T @ x = y
     x = vec6f()
     for i in range(5, -1, -1):
-        s = float(0.0)
+        s = wp.float32(0.0)
         for k in range(i + 1, 6):
             s += L[k, i] * x[k]
         x[i] = (y[i] - s) / L[i, i]
@@ -223,7 +223,7 @@ def lm_step(
     H = mat66f()
     for a in range(6):
         for b in range(6):
-            val = float(0.0)
+            val = wp.float32(0.0)
             for k in range(nv):
                 Ja = jacp[i, a, k] if a < 3 else jacr[i, a - 3, k]
                 Jb = jacp[i, b, k] if b < 3 else jacr[i, b - 3, k]
@@ -237,7 +237,7 @@ def lm_step(
 
     # q_dot = J^T @ x, dq = q_dot * dt
     for k in range(nv):
-        val = float(0.0)
+        val = wp.float32(0.0)
         for a in range(3):
             val += jacp[i, a, k] * x[a]
             val += jacr[i, a, k] * x[a + 3]
@@ -395,7 +395,12 @@ class SimpleWarpKinematics(ParallelKinematics):
 
         qpos_arr = self._dicts_to_qpos_arr(qpos_dicts)
         with wp.ScopedDevice(self._device):
-            wp.copy(data.qpos, wp.from_numpy(qpos_arr))
+            # NOTE: dtype= is required. Without it, wp.from_numpy infers a *vector* dtype for a
+            # 2D array (shape (B, nq) -> B elements of vec{nq}f), whose element size (4*nq bytes)
+            # no longer matches data.qpos's float32 elements. warp <=1.14 only compared total
+            # byte counts so the raw memcpy happened to be correct; warp 1.15 added an element-size
+            # equality check to wp.copy and now raises "Incompatible array data types".
+            wp.copy(data.qpos, wp.from_numpy(qpos_arr, dtype=wp.float32))
             mjw.fwd_position(self._mjw_model, data)
 
         dol = {}
@@ -612,13 +617,17 @@ class SimpleWarpKinematics(ParallelKinematics):
             ik_args.leaf_frame_type.fill_(leaf_frame_type.value)
             ik_args.damping.fill_(damping)
             ik_args.dt.fill_(dt)
+            # NOTE: dtype= is required on both of these; see the comment in forward_kinematics.
             wp.copy(
                 ik_args.jacobian_mask,
-                wp.from_numpy(self._create_jacobian_mask(batch_size, unlocked_move_group_ids)),
+                wp.from_numpy(
+                    self._create_jacobian_mask(batch_size, unlocked_move_group_ids),
+                    dtype=wp.int32,
+                ),
             )
 
             q0_arr = self._dicts_to_qpos_arr(q0_dicts)
-            wp.copy(data.qpos, wp.from_numpy(q0_arr))
+            wp.copy(data.qpos, wp.from_numpy(q0_arr, dtype=wp.float32))
 
             for i in range(max_iter):
                 if self._device.startswith("cuda"):
